Extracted static methods from ScanUtils into a dedicated ConfigurationOverrider service.
