+++ src/main/java/com/checkmarx/flow/controller/GitHubController.java
@@ -64,6 +64,8 @@ public class GitHubController extends WebhookController {
     private final GitHubService gitHubService;
     private final SastScanner sastScanner;
     private final FilterFactory filterFactory;
+    private final ConfigurationOverrider configOverrider;
+
     private Mac hmac;
 
     @PostConstruct
@@ -194,7 +196,7 @@ public class GitHubController extends WebhookController {
 
             /*Check for Config as code (cx.config) and override*/
             CxConfig cxConfig =  gitHubService.getCxConfigOverride(request);
-            request = ScanUtils.overrideCxConfig(request, cxConfig, flowProperties);
+            request = configOverrider.overrideCxConfig(cxConfig, request);
 
             request.putAdditionalMetadata(ScanUtils.WEB_HOOK_PAYLOAD, body);
             request.putAdditionalMetadata("statuses_url", event.getPullRequest().getStatusesUrl());
@@ -310,7 +312,7 @@ public class GitHubController extends WebhookController {
 
             /*Check for Config as code (cx.config) and override*/
             CxConfig cxConfig =  gitHubService.getCxConfigOverride(request);
-            request = ScanUtils.overrideCxConfig(request, cxConfig, flowProperties);
+            request = configOverrider.overrideCxConfig(cxConfig, request);
 
             request.putAdditionalMetadata(ScanUtils.WEB_HOOK_PAYLOAD, body);
             request.setId(uid);
