+++ src/test/java/com/checkmarx/flow/cucumber/integration/azure/publishing/githubflow/PublishingSteps.java
@@ -56,6 +56,8 @@ public class PublishingSteps extends PublishingStepsBase {
     private final ResultsService resultsService;
     private final VulnerabilityScanner sastScanner;
     private final FilterFactory filterFactory;
+    private final ConfigurationOverrider configOverrider;
+
 
     @MockBean
     private final CxClient cxClientMock;
@@ -140,7 +142,7 @@ public class PublishingSteps extends PublishingStepsBase {
         FlowService flowService = new FlowService(vulnerabilityScannerList, projectNameGenerator, resultsService);
 
         return new GitHubController(gitHubProperties, flowProperties, cxProperties,
-                null, flowService, helperService, gitHubService, null, filterFactory);
+                null, flowService, helperService, gitHubService, null, filterFactory, configOverrider);
     }
 
     private static GitHubTestUtils.EventType determineEventType(String eventName) {
