+++ src/test/java/com/checkmarx/flow/cucumber/integration/sast/config/OverwritingProjectConfigSteps.java
@@ -61,6 +61,7 @@ public class OverwritingProjectConfigSteps {
     private final FlowService flowService;
     private final SastScanner sastScanner;
     private final FilterFactory filterFactory;
+    private final ConfigurationOverrider configOverrider;
 
     private Integer projectId;
     private Integer interceptedScanId;
@@ -123,7 +124,7 @@ public class OverwritingProjectConfigSteps {
     public void githubNotifiesCxFlowAboutAPullRequest(String projectName) {
 
         GitHubController gitHubController = new GitHubController(gitHubProperties, flowProperties, cxProperties,
-                null, flowService, helperService, gitHubService, sastScanner, filterFactory);
+                null, flowService, helperService, gitHubService, sastScanner, filterFactory, configOverrider);
 
         gitHubTestUtils.callController(gitHubController, GitHubTestUtils.EventType.PULL_REQUEST, projectName);
     }
